+++ tests/terraform/checks/resource/gcp/example_CloudSqlMajorVersion/main.tf
@@ -94,7 +94,7 @@ resource "google_sql_database_instance" "fail2" {
 }
 
 resource "google_sql_database_instance" "pass2" {
-  database_version = "POSTGRES_15"
+  database_version = "POSTGRES_16"
   name             = "general-pos121"
   project          = "gcp-bridgecrew-deployment"
   region           = "us-central1"